reorganize custom functions; tested terminal+notebook
