+++ prog/run_hero_opt.py
@@ -1,8 +1,6 @@
 ## THIS FILE USES AVAILABLE GEAR TO OPTIMIZE STATS ON YOUR HEROES IN EPIC Seven
 
-## ## FILE SETTINGS
-import setup as st
-import fx_lib as fx
+## ## Import packages
 from datetime import datetime
 import pandas as pd
 pd.set_option('display.max_rows', 100)
@@ -12,23 +10,27 @@ import itertools
 import pickle
 import json
 import yaml
+## ## Import Custom-funtions from python files
+import setup as st
+import fx_lib as fx
+import customer_ui_fx as ui
+## ## Open input datasets
 with open(r'../inp/character_inputs.yaml') as file:
     target_stats = yaml.load(file, Loader=yaml.FullLoader)
 with open('../inp/master_data.json') as json_file:
     data = json.load(json_file)
-##gear combination set up
-import set_combo as sc
-import customer_ui_fx_jn as ui
-
+df_items = pd.read_pickle('../outp/equip_potential.pkl')
+df_items['reco'] = ''
+df_items['start_loc'] = df_items['hero']
 # CATCH INPUT ERRORS
 fx.verify_setup()
 
 # ##### HEROES
-hero_order = sc.startup_msg1(target_stats)
+hero_order = fx.startup_msg1(target_stats)
 ##lock gear on user specified heroes
 lock_gear = target_stats['Lock_Gear']
 df_hero, char_list = fx.hero_json_to_df(hero_order, data)
-sc.startup_msg2(lock_gear)
+df_items = fx.startup_msg2(df_items, lock_gear)
 
 ###input for loop
 for j in range(0,len(hero_order)):
@@ -53,8 +55,8 @@ for j in range(0,len(hero_order)):
 
     run_counter = run_pass = 0
     while (run_pass < 1)&(run_counter < 1):
-        gear_comb_dict = fx.set_combo(sc.df_items[(sc.df_items.set.isin(include_sets))|(sc.df_items.hero==char)], char, sc.l4, sc.l2)  ## Output gear_comb_dict[ [set_nm] , [type] , [ID] ]
-        sc_output = sc.set_combination_iterate(gear_comb_dict, fx.set_4[fx.set_4.Set_Nm.isin(include_sets)].Set_Nm.values , fx.set_2[fx.set_2.Set_Nm.isin(include_sets)].Set_Nm.values, FORCE_4SET = 0)
+        gear_comb_dict = fx.set_combo(df_items[(df_items.set.isin(include_sets))|(df_items.hero==char)], char, fx.l4, fx.l2)  ## Output gear_comb_dict[ [set_nm] , [type] , [ID] ]
+        sc_output = fx.set_combination_iterate(gear_comb_dict, fx.set_4[fx.set_4.Set_Nm.isin(include_sets)].Set_Nm.values , fx.set_2[fx.set_2.Set_Nm.isin(include_sets)].Set_Nm.values, FORCE_4SET = 0)
         print('Progress: Step 1/4 Complete.  Number of combinations found', len(sc_output) , "   ", datetime.now())
         if len(sc_output) == 0:
             print("No combinations were found for this hero. To skip this hero and begin the next here, enter [Skip]. ")
@@ -83,10 +85,10 @@ for j in range(0,len(hero_order)):
     except:
         continue
 
-    sc_df, hero_with_gear = sc.final_gear_combos(sc_output, char)
-    odf = fx.get_combo_stats(sc_df, df_hero, fx.mainst_sum(sc_df, sc.df_items), fx.subst_sum(sc_df, sc.df_items), \
+    sc_df, hero_with_gear = fx.final_gear_combos(sc_output, char)
+    odf = fx.get_combo_stats(sc_df, df_hero, fx.mainst_sum(sc_df, df_items), fx.subst_sum(sc_df, df_items), \
                     fx.set_sum(sc_df), fx.bonus_eqp_sum(df_hero[df_hero.Name == char]), char, hero_target)
-    idx_reco, choice_df = sc.run_stat_reco(odf, hero_with_gear, hero_target)
+    idx_reco, choice_df = fx.run_stat_reco(odf, hero_with_gear, hero_target)
     if st.MANUAL_SELECTION == 1:
         print("Enter the index value (the top number of the stat output) for the stats you would like to assign to your hero:")
         screen_options = choice_df[['Set_1','Set_2','Set_3','WW','ATK','HP','DEF','SPD','CRIT','CDMG','EFF','RES','Dmg_Rating','EHP','Complete']].transpose()
@@ -95,14 +97,14 @@ for j in range(0,len(hero_order)):
         idx_reco = row_idx
     gear_selected = odf.loc[idx_reco]
     odf['Complete'].loc[idx_reco] = 'RECO'
-    sc.df_items = ui.save_hero(sc.df_items, odf.loc[idx_reco], char)
+    df_items = ui.save_hero(df_items, odf.loc[idx_reco], char)
     print("Progress: Step 4/4 Complete.  Recommended gear selected.  Completed hero: ", char)
     if hero_with_gear == 1:  print("Hero began optimization with gear equipped")
     print(odf[odf.Complete.isin(['CURRENT','PREVIOUS','RECO'])][['Complete','Set_1','Set_2','Set_3','WW','ATK','HP','DEF','SPD','CRIT','CDMG','EFF','RES','Dmg_Rating','EHP']])
 
     reco_list = gear_selected['gear_list']
-    sc.df_items['reco'] = np.where( sc.df_items.id.isin(reco_list) , char, sc.df_items['reco'])
+    df_items['reco'] = np.where( df_items.id.isin(reco_list) , char, df_items['reco'])
 ##end for loop
-ui.save_final_data(sc.df_items)
+ui.save_final_data(df_items)
 print("Saving final results.  Results can be viewed in gear_reco.csv or upd_items.json.")
 print("If you are happy with the results, you can move gear over in Epic Seven and update the items section in master_data.json")
